# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mini Web App API",
    "description": "A simple web application with vulnerability testing endpoints built with Node.js and Express",
    "version": "1.0.0",
    "contact": {
      "name": "Mini Web App",
      "url": "http://localhost:3000"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://your-domain.com",
      "description": "Production server (example)"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Home page",
        "description": "Serves the main HTML page",
        "responses": {
          "200": {
            "description": "HTML page served successfully",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Page not found",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/about": {
      "get": {
        "summary": "About page",
        "description": "Serves the about HTML page",
        "responses": {
          "200": {
            "description": "About page served successfully",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Page not found"
          }
        }
      }
    },
    "/contact": {
      "get": {
        "summary": "Contact page",
        "description": "Serves the contact HTML page",
        "responses": {
          "200": {
            "description": "Contact page served successfully",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Page not found"
          }
        }
      }
    },
    "/yourpage.html": {
      "get": {
        "summary": "Vulnerability testing page",
        "description": "Serves the vulnerability testing lab page",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Parameter for reflected XSS testing (VULN-1)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "item",
            "in": "query",
            "description": "Parameter for list injection testing (VULN-3)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "link",
            "in": "query",
            "description": "Parameter for link injection testing (VULN-5)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "onclickCode",
            "in": "query",
            "description": "Parameter for onclick injection testing (VULN-7)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eval",
            "in": "query",
            "description": "Parameter for eval injection testing (VULN-8)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeout",
            "in": "query",
            "description": "Parameter for timeout injection testing (VULN-9)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "title",
            "in": "query",
            "description": "Parameter for template injection testing (VULN-11)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "iframeDoc",
            "in": "query",
            "description": "Parameter for iframe srcdoc injection testing (VULN-6)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "write",
            "in": "query",
            "description": "Parameter for document.write injection testing (VULN-4)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vulnerability testing page served successfully",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Page not found"
          }
        }
      }
    },
        "/api/hello": {
          "get": {
            "summary": "Hello API endpoint",
            "description": "Returns a greeting message with timestamp",
            "tags": ["API"],
            "responses": {
              "200": {
                "description": "Successful response",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Hello from the server!"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-01-15T10:30:00.000Z"
                        }
                      },
                      "required": ["message", "timestamp"]
                    }
                  }
                }
              },
              "400": {
                "description": "Bad request",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/Error"
                    }
                  }
                }
              },
              "404": {
                "description": "Not found",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/Error"
                    }
                  }
                }
              },
              "405": {
                "description": "Method not allowed",
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string",
                          "example": "Method not allowed"
                        },
                        "success": {
                          "type": "boolean",
                          "example": false
                        },
                        "allowedMethods": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "example": ["GET"]
                        }
                      },
                      "required": ["error", "success"]
                    }
                  }
                }
              },
              "500": {
                "description": "Internal server error",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/Error"
                    }
                  }
                }
              }
            }
          }
        },
    "/api/contact": {
      "post": {
        "summary": "Contact form submission",
        "description": "Handles contact form submissions",
        "tags": ["API"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Thank you for your message, John! We will get back to you soon."
                    }
                  },
                  "required": ["success", "message"]
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error or invalid JSON",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "msg": {
                            "type": "string"
                          },
                          "param": {
                            "type": "string"
                          },
                          "location": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "success": {
                      "type": "boolean",
                      "example": false
                    }
                  },
                  "required": ["error", "success"]
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/unsafe-html": {
      "get": {
        "summary": "Unsafe HTML endpoint (for testing)",
        "description": "Returns HTML content for vulnerability testing (VULN-10)",
        "tags": ["Testing"],
        "responses": {
          "200": {
            "description": "HTML content returned",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Endpoint not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ContactForm": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User's name",
            "example": "John Doe",
            "minLength": 1,
            "maxLength": 100
          },
          "message": {
            "type": "string",
            "description": "User's message",
            "example": "Hello, I have a question about your service.",
            "minLength": 1,
            "maxLength": 1000
          }
        },
        "required": ["name", "message"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "An error occurred"
          },
          "success": {
            "type": "boolean",
            "example": false
          }
        },
        "required": ["error"]
      }
    },
    "securitySchemes": {}
  },
  "security": [],
  "tags": [
    {
      "name": "API",
      "description": "Main API endpoints"
    },
    {
      "name": "Testing",
      "description": "Vulnerability testing endpoints"
    },
    {
      "name": "Pages",
      "description": "Static page endpoints"
    }
  ]
}